+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Bug fixes:
 Compatibility:
 
 * Implement `Process::Status.wait` (#2378).
+* Update `rb_str_modify` and `rb_str_modify_expand` to raise a `FrozenError` when given a frozen string (#2392).
 
 Performance: